+++ compiler/parser/lexer.h
@@ -26,6 +26,7 @@ int skip_comment(void);
 Token* read_string(void);
 Token* read_number(void);
 Token* read_identifier(void);
+Token* read_raw_identifier(void);
 
 #endif // LEXER_H